chore(tests): separate test data from test files
- Create folder 'testData' inside the server folder
- Create test data files for tests
- Move test data from test files to test data files inside testData folder
- Import test data into test files

[Finishes #165376546]
