+++ server/testData/accounts.js
@@ -1,0 +1,59 @@
+const testAccount = {
+  accountNumber: 1012579609,
+  owner: 1,
+  firstName: 'Alice',
+  lastName: 'Nwankwo',
+  email: 'alicen1995@yahoo.com',
+  type: 'savings',
+  openingBalance: 120350.05,
+  status: 'active',
+};
+
+const testAccountData = {
+  accountNumber: 1024857699,
+  owner: 1,
+  firstName: 'Alice',
+  lastName: 'Nwankwo',
+  email: 'alicen1995@yahoo.com',
+  type: 'current',
+  openingBalance: 45000.35,
+  status: 'dormant',
+};
+const testTransactionAccountData = {
+  accountNumber: 1025729475,
+  owner: 1,
+  firstName: 'Alice',
+  lastName: 'Nwankwo',
+  email: 'alicen1995@yahoo.com',
+  type: 'current',
+  openingBalance: 45000.35,
+  status: 'dormant',
+};
+const delTestAccountData = {
+  accountNumber: 1024545001,
+  owner: 1,
+  firstName: 'Alice',
+  lastName: 'Nwankwo',
+  email: 'alicen1995@yahoo.com',
+  type: 'current',
+  openingBalance: 45000.35,
+  status: 'dormant',
+};
+const getTestAccountData = {
+  accountNumber: 1025887203,
+  owner: 1,
+  firstName: 'Alice',
+  lastName: 'Nwankwo',
+  email: 'alicen1995@yahoo.com',
+  type: 'current',
+  openingBalance: 45000.35,
+  status: 'dormant',
+};
+
+export default {
+  testAccount,
+  testAccountData,
+  delTestAccountData,
+  getTestAccountData,
+  testTransactionAccountData,
+};
