+++ server/testData/transactions.js
@@ -1,0 +1,26 @@
+const testCreditTransactionData = {
+  type: 'credit',
+  accountNumber: 1024857699,
+  cashier: 1,
+  amount: 3400.00,
+};
+
+const testDebitTransactionData = {
+  type: 'debit',
+  accountNumber: 1024857699,
+  cashier: 1,
+  amount: 3400.00,
+};
+
+const testDummyTransactionData = {
+  type: 'credit',
+  accountNumber: 1024857699,
+  cashier: 1,
+  amount: 300.00,
+};
+
+export default {
+  testCreditTransactionData,
+  testDebitTransactionData,
+  testDummyTransactionData,
+};
