+++ server/testData/users.js
@@ -1,0 +1,21 @@
+const testUser = {
+  firstName: 'Dora',
+  lastName: 'Ofili',
+  email: 'doraofili@yahoo.com',
+  password: 'dorasecurepassword',
+  type: 'client',
+  isAdmin: false,
+};
+const testUserData = {
+  firstName: 'Maureen',
+  lastName: 'George',
+  email: 'maumau@yahoo.com',
+  password: 'maureensecurepassword',
+  type: 'staff',
+  isAdmin: true,
+};
+
+export default {
+  testUser,
+  testUserData,
+};
